% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_crt.R
\name{sim_crt}
\alias{sim_crt}
\title{Simulate a basic cluster randomised trial}
\usage{
sim_crt(
  n_clusters = 10,
  n_cluster_individuals = 100,
  intercept = 0.2,
  trt_effect = 2,
  s2b = 0.05,
  s2w = 0.95
)
}
\arguments{
\item{n_clusters}{Number of cluster in the trial}

\item{n_cluster_individuals}{average number of individuals per cluster}

\item{intercept}{the value of the outcome in the control group}

\item{trt_effect}{the treatment effect in the trial}

\item{s2b}{The between cluster standard devation}

\item{s2w}{The within cluster standard devation}
}
\value{
Data.frame with cluster_id, ind_id, trt, outcome.
}
\description{
This function simulates CRTs for a range of cluster and participants
}
